Improve validation of requiredQualifierConstraint.
Now has the option to check for more than one required qualifier at
once. In this tool we only work with property-constraint pairs, so if a
certain constraint occurs multiple times for the same property, we
implement this as a single constraint of that type combining all of the
arguments. This could be an issue later if we start implementing scopes.