Correct implementation of separator.
In SingleValueConstraint, the usage of separator was wrongly
interpreted. Previous version would consider values different from the
moment a separator was present, even if both values also had the same
value for the separator. Now the validation also compaires separator
value(s). The only slight problem still lingering, is that currently the
presence of a separator on one value while it is abscent on the other,
will also be detected as a different value. It is clear that this
visually different, but not the intended way to use a separator.